requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.340
langchain-core==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.3.0
pydantic==2.5.0
langdetect==1.0.9
pypdf==3.17.1
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
